requirements.txt
torch==1.12.1
torchvision==0.13.1
matplotlib==3.5.3
scikit-learn==1.1.2
pandas==1.5.0
seaborn==0.12.0
tqdm==4.64.1
pyyaml==6.0